Make sure async disposal tests are executed on .NET Framework
